requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-community
faiss-cpu
pypdf
python-docx
gradio
requests
beautifulsoup4